requirements.txt
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple

torch>=2.5.1
torchvision>=0.20.1
accelerate>=1.2.0
torchmetrics>=1.6.0
numpy~=2.3.1
hydra-core>=1.3.2
omegaconf~=2.3.0
pydantic>=2.10.3
tqdm>=4.67.1
wandb>=0.18.7